CRUDUsers_03

+++ src/controllers/users_controller.js
@@ -5,13 +5,14 @@ import { createToken } from '../middlewares/auth.js';
 const saltRounds = 10
 const registerUserController = async(req, res)=> {
     
-    const { password, email, username, ...otherData } = req.body;
+    const { password, email, username, rol, ...otherData } = req.body;
 
 	var validEmail = /^\w+([.-_+]?\w+)*@\w+([.-]?\w+)*(\.\w{2,10})+$/;
 
-	if (!password || !email || !username || password.trim() === "" || email.trim() === "" || username.trim() === "") {
+	if (!password || !email || !username || !rol || password.trim() === "" || email.trim() === "" || username.trim() === "" || rol.trim() === "" ) {
         return res.status(400).json({error: "Datos vacíos, llénelos por favor" });
     }
+
     if( !validEmail.test(email)){
 		res.status(500).json({error: "Email no valido "});
 	}
@@ -26,18 +27,19 @@ const registerUserController = async(req, res)=> {
             password: hashedPassword,
             email, 
             username, 
+            rol, 
             ...otherData
     }
     try {
         const user = await userModel.registerUserModel(userData);
-        console.log(user);
-            
+        if (user.error){
+            return res.status(401).json({error: user.error})
+        }       
         res.status(200).json(user);
     } catch (error) {
         res.status(500).json(error.msg);
     }
 }
-
 
 const loginUserController = async (req, res) => {
     const { username, password } = req.body;
@@ -55,22 +57,32 @@ const loginUserController = async (req, res) => {
 };
 
 const allUsersController = async (req, res) => {
-    try {
-        const allUsers = await userModel.getUsers();
-        res.status(200).json(allUsers);
-    } catch (error) {
-        res.status(500).json({message:error});
+    const {rol} = req.body;
+    if (rol == "administrador" || rol == "admin"){
+        try {
+            const allUsers = await userModel.getUsers();
+            res.status(200).json(allUsers);
+        } catch (error) {
+            res.status(500).json({message:error});
+        }
+    } else{
+        res.status(401).json({error: "Mostrar a todos los usuarios solo se encuentra permitido para el administrador"});
     }
 }
 
 const oneUserController = async (req, res) => {
     const {id} = req.params;
-    try{
-        const oneUser = await userModel.findUser(id);
-        res.status(200).json(oneUser);
-    } catch (error)
-    {
-        res.status(500).json({message:error});
+    const {rol} = req.body;
+    if(rol == "administrador" || rol == "admin"){
+        try{
+            const oneUser = await userModel.findUser(id);
+            res.status(200).json(oneUser);
+        } catch (error)
+        {
+            res.status(500).json({message:error});
+        }
+    } else{
+        res.status(401).json({error: "Mostrar a un usuario en especifico solo se encuentra permitido para el administrador"});
     }
 }
 
@@ -90,15 +102,18 @@ const updateUserController = async (req,res) => {
     }
 }
 
-
 const deleteUserController = async (req,res) => {
     const {id} = req.params
-
-    try {
-        const deleteUser = await userModel.deleteUser(id);
-        res.status(200).json({message:'Se ha eliminado correctamente el usuario'})
-    } catch (error) {
-        res.status(500).json({message:error})
+    const {rol} = req.body;
+    if(rol == "administrador" || rol == "admin"){
+        try {
+            const deleteUser = await userModel.deleteUser(id);
+            res.status(200).json({message:'Se ha eliminado correctamente el usuario'})
+        } catch (error) {
+            res.status(500).json({message:error})
+        }
+    }else{
+        res.status(401).json({error:"Eliminar a un usuario solo se encuentra permitido para el administrador"})
     }
 }
